decouple api call
